CAuth: Make sure we re-create canaries
This starts to fix an issue whereby the canaries were not correctly
re-created on check, but instead could be passed via callee-saved
registers which might go under attacker control.

This approach defines new intrinsics for creating the modifier in the
prologue and epilogue. Because they are different (although they result in
the same value after lowering), the compiler is forced to re-create the
canary.

TODO: This is not yet properly tested!